Added BaseItem as a new abstract class for CraftingSystem as replacement
