+++ Assets/CraftingSystem/Core/BaseItem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1eae31a75464e946a9e25143786a764
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
