+++ Assets/CraftingSystem/Core/RecipeSO.cs
@@ -1,18 +1,70 @@
 using System.Collections;
 using System.Collections.Generic;
+using CraftingSystem.Core;
 using UnityEngine;
 
-public class RecipeSO : MonoBehaviour
+
+//Attribute which allows right click->Create
+[CreateAssetMenu(fileName = "New Recipe", menuName = "Items/New Recipe")]
+public class RecipeSO : ScriptableObject
 {
-    // Start is called before the first frame update
-    void Start()
-    {
+    [SerializeField] public BaseItem _ResultItem;
+    [SerializeField] public int _ResultCount = 1;
+
+    [SerializeField] public Vector2Int _SizeOfGrid;
+    [SerializeField] public BaseItem[] ingredients;
         
+    //Used in game logic
+    private Recipe _recipe;
+        
+    // Used in editor
+    private bool _isInitialized = false;
+    private bool _isRecipeValid = false;
+
+    // Used in editor
+    public Recipe Recipe
+    {
+        get
+        {
+            if (!Application.isEditor && _isInitialized) return _recipe;
+            CreateRecipe();
+            _isInitialized = true;
+
+            return _recipe;
+        }
     }
 
-    // Update is called once per frame
-    void Update()
+    // Validating the recipe
+    private bool IsValid
     {
-        
+        get
+        {
+            if (!Application.isEditor && _isInitialized) return _isRecipeValid;
+            CreateRecipe();
+            _isInitialized = true;
+            return _isRecipeValid;
+        }
+    }
+    
+    // Creating a CreateRecipe() method
+    private void CreateRecipe()
+    {
+        if (_SizeOfGrid == Vector2Int.zero)
+        {
+            _isRecipeValid = false;
+            return;
+        }
+
+        _isRecipeValid = false;
+        // check if at least one item is not null
+        foreach (var item in ingredients)
+        {
+            if (item != null)
+            {
+                _recipe = new Recipe(_SizeOfGrid, ingredients, _ResultItem, _ResultCount);
+                _isRecipeValid = _ResultCount > 0;
+                return;
+            }
+        }
     }
 }
